Fix checkmc command

+++ src/test/java/seedu/address/testutil/TypicalModules.java
@@ -51,10 +51,10 @@ public class TypicalModules {
         return new ArrayList<>(Arrays.asList(CS2103T, CS3216));
     }
 
-    public static int getTypicalTotalMc() {
-        int totalMc = 0;
+    public static double getTypicalTotalMc() {
+        double totalMc = 0;
         for (Module module: getTypicalModules()) {
-            totalMc += Integer.parseInt(module.getModularCredits().toString());
+            totalMc += Double.parseDouble(module.getModularCredits().toString());
         }
         return totalMc;
     }
